Initial release (#1)
I've been writing up a blog post and leveraging KaTeX via MathEmbed/screenshots. Unfortunately, this is hard to reuse due to link copy/pasting. This came to a head after I found an error in my proofs so I likely need to rewrite a lot of formulas

As a result of these 2 issues, I've decided to write up a screenshotting tool for KaTeX. In this PR:

- Initial release for `katex-screenshot`
- Added documentation
- Added library
- Added tests